requirements.txt
fastapi
pydantic
python-dotenv
requests
openai==1.84.0
hypercorn
python-multipart
aiofiles
pytest-asyncio

# Database dependencies
sqlalchemy[asyncio]==2.0.*
psycopg2-binary==2.9.*
asyncpg==0.29.*
alembic==1.13.*

motor==3.3.1

# Firebase dependencies
firebase_admin==6.2.0